pyproject.toml
[tool.poetry]
name = "dex-agg-tutorial"
version = "0.1.0"
description = ""
authors = ["ertemann <info@crypticcurve.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruff = "^0.12.8"
black = "^25.1.0"
mypy = "^1.17.1"
djangorestframework = "^3.16.1"
django = "^5.2.5"
web3 = "^7.13.0"
retry = "^0.9.2"
python-dotenv = "^1.1.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
